zzip: fseeko: fix comparison between incompatible pointer types
The C standard does not allow for comparing incompatible pointer types
with each other. This causes the C compiler to emit a warning when we
try to compare a pointer of type `struct zzip_disk64_trailer` (cast to
`void *`) with a pointer of type `unsigned char`.

Fix this by directly comparing `p + 1` with `buffer + mapsize` instead
of using the cast struct.