cmake: do not modify test file in source dir
When doing out-of-tree builds, a golden rule is to not touch any files
inside of the source directory. This rule is being violated by our tests
CMake instructions, which generate a "test0.zip" file and copy it over
the file "test/test.zip" in our source directory.

Stop doing this. We already copy the "test0.zip" to "test.zip" in the
binary directory, and all tests should be using that one.