test: urllib compatibility with Python 3
In Python 3, the functions `quote_plus` and `urlretrieve` have been
moved from urllib to urllib.parse and urllib.request, respectively. To
provide compatibility with the old and new module layouts, use a
try-catch block and fall back to the new layout if an ImportError was
raised.

Furthermore, do not hand-code `urlretrieve` functionality with
`urlopen` to avoid additional compatibility issues.