test: specify subprocess encoding to be UTF-8
One of the most important and user-visible changes in Python 3 was that
when reading strings from some , one always needs to specify the
encoding lest the dreaded UnicodeException would occur if there were any
characters outside of the ASCII range. In the test suite, we're using a
subprocess taht communicates with the shell, but we do not specify the
encoding of both its stdout and stderr previous to reading them.

Explicitly decode both their output and error streams to UTF-8 to avoid
any exceptions raised by Python 3.